+++ packages/common/package.json
@@ -5,7 +5,9 @@
     "mongoose": "^7.0.2",
     "winston": "^3.10.0",
     "winston-mongodb": "^5.1.1",
-    "axios": "^1.5.1"
+    "axios": "^1.5.1",
+    "nextjs-cors": "^2.1.2",
+    "next": "13.5.2"
   },
   "devDependencies": {
     "@types/node": "^20.8.5"
