+++ packages/common/tools/stream.ts
