+++ projects/app/src/pages/api/openapi/v1/chat/completions.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { withNextCors } from '@/service/utils/tools';
+import { withNextCors } from '@fastgpt/common/tools/nextjs';
 import ChatCompletion from '@/pages/api/v1/chat/completions';
 
 export default withNextCors(async function handler(req: NextApiRequest, res: NextApiResponse) {
