+++ projects/app/src/pages/api/user/account/loginByPassword.ts
@@ -2,8 +2,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { jsonRes } from '@/service/response';
 import { MongoUser } from '@fastgpt/support/user/schema';
-import { setCookie } from '@/service/utils/tools';
-import { generateToken } from '@fastgpt/support/user/tools';
+import { setCookie } from '@fastgpt/support/user/auth';
+import { generateToken } from '@fastgpt/support/user/auth';
 import { connectToDatabase } from '@/service/mongo';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
