+++ projects/app/src/service/utils/tools.ts
@@ -1,36 +1,6 @@
-import type { NextApiResponse, NextApiHandler, NextApiRequest } from 'next';
-import NextCors from 'nextjs-cors';
+import type { NextApiResponse } from 'next';
 import { generateQA } from '../events/generateQA';
 import { generateVector } from '../events/generateVector';
-
-/* set cookie */
-export const setCookie = (res: NextApiResponse, token: string) => {
-  res.setHeader(
-    'Set-Cookie',
-    `token=${token}; Path=/; HttpOnly; Max-Age=604800; Samesite=None; Secure;`
-  );
-};
-/* clear cookie */
-export const clearCookie = (res: NextApiResponse) => {
-  res.setHeader('Set-Cookie', 'token=; Path=/; Max-Age=0');
-};
-
-export function withNextCors(handler: NextApiHandler): NextApiHandler {
-  return async function nextApiHandlerWrappedWithNextCors(
-    req: NextApiRequest,
-    res: NextApiResponse
-  ) {
-    const methods = ['GET', 'eHEAD', 'PUT', 'PATCH', 'POST', 'DELETE'];
-    const origin = req.headers.origin;
-    await NextCors(req, res, {
-      methods,
-      origin: origin,
-      optionsSuccessStatus: 200
-    });
-
-    return handler(req, res);
-  };
-}
 
 /* start task */
 export const startQueue = () => {
@@ -41,20 +11,6 @@ export const startQueue = () => {
   for (let i = 0; i < global.systemEnv.vectorMaxProcess; i++) {
     generateVector();
   }
-};
-
-export const sseResponse = ({
-  res,
-  event,
-  data
-}: {
-  res: NextApiResponse;
-  event?: string;
-  data: string;
-}) => {
-  if (res.closed) return;
-  event && res.write(`event: ${event}\n`);
-  res.write(`data: ${data}\n\n`);
 };
 
 /* add logger */
